pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xsoar-command-toolkit"
version = "0.1.0"
description = "CLI toolkit to interact with Cortex XSOAR via REST API"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Jersé Nunes" }
]
dependencies = [
  "requests>=2.32.5",
  "python-dotenv>=1.2.1"
]

[project.scripts]
xsoar-toolkit = "xsoar_command_toolkit.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
